Add hakyll unicode support
ondrej and I found that in pure environments like a nix build, `$LANG`
gets unset, which hakyll needs for Unicode support. Adding this to the
derivation and shell seems to work against sample hakyll projects.

Setting this up preemptively so that I don't have to worry about it in
the future.